+++ packages/svelte/src/constants.js
@@ -273,30 +273,3 @@ export function is_tag_valid_with_parent(tag, parent_tag) {
 
 	return true;
 }
-
-const ATTR_REGEX = /["<]/g;
-const CONTENT_REGEX = /[&<]/g;
-
-/**
- * @template V
- * @param {V} value
- * @param {boolean} [is_attr]
- */
-export function escape_html(value, is_attr) {
-	const str = String(value ?? '');
-
-	const pattern = is_attr ? ATTR_REGEX : CONTENT_REGEX;
-	pattern.lastIndex = 0;
-
-	let escaped = '';
-	let last = 0;
-
-	while (pattern.test(str)) {
-		const i = pattern.lastIndex - 1;
-		const ch = str[i];
-		escaped += str.substring(last, i) + (ch === '&' ? '&amp;' : ch === '"' ? '&quot;' : '&lt;');
-		last = i + 1;
-	}
-
-	return escaped + str.substring(last);
-}
